requirements.txt
requests
python-dotenv
pytz
openai
alpaca-trade-api
